+++ lib/iap_verifier.rb
@@ -1,7 +1,6 @@
 require "iap_verifier/version"
 require 'iap_verifier/error'
 require 'iap_verifier/request'
-require 'iap_verifier/request_data'
 require 'iap_verifier/response_data'
 require 'iap_verifier/receipt'