+++ lib/iap_verifier/receipt.rb
@@ -3,10 +3,6 @@ module IAPVerifier
     def initialize(response_data)
       @receipt_data = response_data['receipt']
 
-      unless @receipt_data.is_a?(Hash)
-        raise Error::MalformedReceiptData
-      end
-
       @receipt_data.each do |key, value|
         define_singleton_method key.to_s, -> { value }
       end