+++ test/response_data_test.rb
@@ -5,14 +5,14 @@ require 'iap_verifier/error'
 
 class ResponseDataTest < MiniTest::Test
   def test_success_response_json
-    success_response_json = { 'status' => 0, 'receipt' => { } }
+    success_response_json = { 'status' => 0, 'receipt' => { } }.to_json
     response_data = IAPVerifier::ResponseData.new(success_response_json)
 
     assert_equal true, response_data.valid?
   end
 
   def test_error_response_json
-    error_response_json = { 'status' => 21002 }
+    error_response_json = { 'status' => 21002 }.to_json
     response_data = IAPVerifier::ResponseData.new(error_response_json)
 
     assert_equal false, response_data.valid?
@@ -20,7 +20,7 @@ class ResponseDataTest < MiniTest::Test
 
   def test_sandbox_response_json
     # sandbox receipt sent to production server
-    sandbox_response_json = { 'status' => 21007 }
+    sandbox_response_json = { 'status' => 21007 }.to_json
     response_data = IAPVerifier::ResponseData.new(sandbox_response_json)
 
     assert_equal false, response_data.valid?
@@ -28,17 +28,22 @@ class ResponseDataTest < MiniTest::Test
   end
 
   def test_receipt
-    response_data = IAPVerifier::ResponseData.new({ 'receipt' => { } })
+    response_data = IAPVerifier::ResponseData.new({ 'receipt' => { } }.to_json)
     response_data.stubs(:valid?).returns(true)
 
     assert_equal true, response_data.receipt.instance_of?(IAPVerifier::Receipt)
   end
 
-  def test_invalid_response_data
-    assert_raises(IAPVerifier::Error::InvalidResponseData) do
-      response_data = IAPVerifier::ResponseData.new({ 'receipt' => { } })
-      response_data.stubs(:valid?).returns(false)
+  def test_malformed_response_data
+    assert_raises(IAPVerifier::Error::MalformedResponseData) do
+      IAPVerifier::ResponseData.new('')
+    end
+  end
 
+  def text_invalid_receipt_data
+    assert_raises(IAPVerifier::Error::InvalidReceiptData) do
+      IAPVerifier::ResponseData.new({ 'receipt' => { } }.to_json)
+      response_data.stubs(:valid?).returns(false)
       response_data.receipt
     end
   end